ANDROID: Cast time64 values to long long
Resolves overflow compiler warnings from clang.

Test: mma

+++ tools/testing/selftests/timers/nsleep-lat.c
@@ -146,7 +146,7 @@ int main(int argc, char **argv)
 			continue;
 
 		length = 10;
-		while (length <= (NSEC_PER_SEC * 10)) {
+		while (length <= ((long long) NSEC_PER_SEC * 10)) {
 			ret = nanosleep_lat_test(clockid, length);
 			if (ret)
 				break;
